+++ pkg/tracing/tracing.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/go-kit/log/level"
 	"github.com/pkg/errors"
+	"github.com/sercand/kuberesolver"
 	"github.com/weaveworks/common/tracing"
 	"go.opentelemetry.io/contrib/propagators/aws/xray"
 	"google.golang.org/grpc/credentials"
@@ -41,6 +42,7 @@ type Otel struct {
 	OtlpEndpoint   string              `yaml:"otlp_endpoint" json:"otlp_endpoint"`
 	ExporterType   string              `yaml:"exporter_type" json:"exporter_type"`
 	SampleRatio    float64             `yaml:"sample_ratio" json:"sample_ratio"`
+	RoundRobin     bool                `yaml:"round_robin" json:"round_robin"`
 	TLSEnabled     bool                `yaml:"tls_enabled"`
 	TLS            tls.ClientConfig    `yaml:"tls"`
 	ExtraDetectors []resource.Detector `yaml:"-"`
@@ -55,6 +57,7 @@ func (c *Config) RegisterFlags(f *flag.FlagSet) {
 	f.StringVar(&c.Otel.OtlpEndpoint, p+".otel.otlp-endpoint", "", "otl collector endpoint that the driver will use to send spans.")
 	f.StringVar(&c.Otel.ExporterType, p+".otel.exporter-type", "", "enhance/modify traces/propagators for specific exporter. If empty, OTEL defaults will apply. Supported values are: `awsxray.`")
 	f.BoolVar(&c.Otel.TLSEnabled, p+".otel.tls-enabled", c.Otel.TLSEnabled, "Enable TLS in the GRPC client. This flag needs to be enabled when any other TLS flag is set. If set to false, insecure connection to gRPC server will be used.")
+	f.BoolVar(&c.Otel.RoundRobin, p+".otel.round-robin", false, "If enabled, use round_robin gRPC load balancing policy. By default, use pick_first policy. For more details, please refer to https://github.com/grpc/grpc/blob/master/doc/load-balancing.md#load-balancing-policies.")
 	c.Otel.TLS.RegisterFlagsWithPrefix(p+".otel.tls", f)
 }
 
@@ -91,15 +94,23 @@ func SetupTracing(ctx context.Context, name string, c Config) (func(context.Cont
 			level.Warn(util_log.Logger).Log("msg", "DEPRECATED: otel.otlp and otel.oltp both set, using otel.otlp because otel.oltp is deprecated")
 		}
 
-		options := []otlptracegrpc.Option{
-			otlptracegrpc.WithEndpoint(c.Otel.OtlpEndpoint),
-		}
-
+		endpoint := c.Otel.OtlpEndpoint
 		if (c.Otel.OtlpEndpoint == "") && (len(c.Otel.OltpEndpoint) > 0) {
 			level.Warn(util_log.Logger).Log("msg", "DEPRECATED: otel.oltp is deprecated use otel.otlp")
-			options = []otlptracegrpc.Option{
-				otlptracegrpc.WithEndpoint(c.Otel.OltpEndpoint),
-			}
+			endpoint = c.Otel.OltpEndpoint
+		}
+		options := []otlptracegrpc.Option{
+			otlptracegrpc.WithEndpoint(endpoint),
+		}
+		// Following https://github.com/sercand/kuberesolver/blob/master/builder.go#L96.
+		if strings.HasPrefix(endpoint, "kubernetes://") {
+			// Registers the kuberesolver which resolves endpoint with prefix kubernetes://
+			// as kubernetes service endpoint addresses.
+			kuberesolver.RegisterInCluster()
+		}
+
+		if c.Otel.RoundRobin {
+			options = append(options, otlptracegrpc.WithServiceConfig(`{"loadBalancingPolicy": "round_robin"}`))
 		}
 
 		if c.Otel.TLSEnabled {